requirements/common.txt
# Common dependencies for all project environments.

argon2_cffi==18.1.0
brotlipy==0.7.0
Django==2.1.2
django-allauth==0.36.0
django-anymail[sendgrid]==3.0
django-configurations==2.0
django-cors-headers==2.2.0
django-environ==0.4.4
django-extensions==2.0.7
django-log-request-id==1.3.2
django-mail-templated==2.6.5
django-sites==0.10
django-storages==1.6.6
django-uuid-upload-path==1.0.0
django-versatileimagefield==1.9
gunicorn==19.8.1
mkdocs==0.17.4
Pillow==5.0.0
pytz==2018.3
raven==6.9.0
whitenoise==3.3.1
